+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Improvements:
 * [OLMIS-3563:](https://openlmis.atlassian.net/browse/OLMIS-3563) Added moving ng-if from input to wrapper element.
 * [OLMIS-3080:](https://openlmis.atlassian.net/browse/OLMIS-3080) Removed warning modal.
 * [OLMIS-2695](https://openlmis.atlassian.net/browse/OLMIS-2695): Added 'add days to date' method to date utils.
+* [OLMIS-2695](https://openlmis.atlassian.net/browse/OLMIS-2695): Added disabled attribute to datepicker.
 
 5.2.0 / 2017-11-09
 ==================